+++ apps/Common/include/BC4DepthVideoTexture.h
@@ -33,7 +33,7 @@ public:
 
     BC4DepthVideoTexture(const TextureDataCreateParams& params, std::string streamerURL);
 
-    void setMaxQueueSize(uint maxQueueSize) {
+    void setMaxQueueSize(size_t maxQueueSize) {
         this->maxQueueSize = maxQueueSize;
     }
 
@@ -48,7 +48,7 @@ private:
     time_t prevTime = timeutils::getTimeMicros();
     pose_id_t prevPoseID = -1;
 
-    uint maxQueueSize = 10;
+    size_t maxQueueSize = 10;
     std::mutex m;
 
     struct FrameData {
@@ -56,7 +56,7 @@ private:
         std::vector<char> buffer;
     };
     std::deque<FrameData> depthFrames;
-    uint compressedSize;
+    size_t compressedSize;
 
     ZSTDCodec codec;
